add endpoint for file uploads - create new function for handling uploads - add pre-commit hook for secrets
